requirements.txt
mysql-connector-python==8.0.27
Flask==2.0.2
PyJWT==1.7.1
schedule~=1.1.0
Werkzeug~=2.0.2